# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger MMS",
    "description": "locpro's magic",
    "termsOfService": "",
    "contact": {
      "email": "nhoxloctran125@gmail.com"
    },
    "license": {
      "name": "No license",
      "url": ""
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "localhost"
    },
    {
      "url": "https://capstone-api-mms.herokuapp.com/",
      "description": "Heroku app"
    }
  ],
...
  "paths": {
...
    "/users/sign-out": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Sign out the current user",
        "operationId": "signOut",
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "example": " "
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Error: Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get the current user",
        "operationId": "getCurrentUser",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "example": " "
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Error: Internal Server Error"
          }
        }
      }
    },
    "/users/logout-all": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Sign out all user",
        "operationId": "signOutAll",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "example": " "
              }
            }
          },
          "405": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Error: Internal Server Error"
          }
        }
      }
    },
...
    "/api/sendo/attribute/{attributeId}": {
      "get": {
        "tags": [
          "sendoAPI"
        ],
        "summary": "Get attribute by attributeId",
        "operationId": "getAttribute",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "attributeId",
            "required": true,
            "description": "Numeric ID of attribute to get",
            "schema": {
              "type": "integer",
              "example": 1485
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
    "/api/lazada/attribute/{categoryId}": {
      "get": {
        "tags": [
          "lazadaAPI"
        ],
        "summary": "get Attributes by categoryId on lazada",
        "operationId": "getAttributes",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "number",
              "example": 10100183
            },
            "description": "categoryId must be primary category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Attributes not found",
            "content": {
              "application/json": {
                "example": {
                  "code": "601",
                  "type": "ISP",
                  "message": "E601: Requested category not exists",
                  "request_id": "0b1195dd16111364977805848"
                }
              }
            }
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Platform-Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}